+++ client/src/components/ProtectedRoute.tsx
@@ -1,16 +1,19 @@
-// src/components/ProtectedRoute.jsx
 import { Navigate, useLocation } from 'react-router-dom';
+import { ReactNode } from 'react';
 
-const ProtectedRoute = ({ children }) => {
+interface ProtectedRouteProps {
+  children: ReactNode;
+}
+
+const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
   const token = localStorage.getItem('token');
   const location = useLocation();
-  
+
   if (!token) {
-    // Redirect to login if not authenticated
     return <Navigate to="/login" state={{ from: location }} replace />;
   }
-  
-  return children;
+
+  return <>{children}</>;
 };
 
 export default ProtectedRoute;
